do not swallow errors

+++ src/core/server/logging/integration_tests/rolling_file_appender.test.ts
@@ -48,12 +48,10 @@ describe('RollingFileAppender', () => {
   });
 
   afterEach(async () => {
-    try {
+    if (testDir) {
       await rmdir(testDir, { recursive: true });
-    } catch (e) {
-      // eslint-disable-next-line no-console
-      console.error(e);
     }
+
     if (root) {
       await root.shutdown();
     }
